pyproject.toml
[project]
name = "testknit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "pyobjc>=12.1",
    "tkmacosx>=1.0.5",
    "tkmacosx; sys_platform == 'darwin'",
]
